Работа с SQL [~] Немного изменил синтаксиси и либы sqldelight_scheme.sq. [-] Удалил не нужные либы из gradle. [+] Теперь при окончании тестирования данные о тесте буду записываться в SQL таблицу Важно! В этот коммите собирается exe файл.

+++ src/main/kotlin/domain/model/regular/tests/TestMode.kt
@@ -2,7 +2,7 @@ package domain.model.regular.tests
 
 data class TestMode(
     val name: String?,
-    val id: Int?
+    val id: Long?
 ) {
     companion object {
         fun empty(): TestMode {
